Verify hub before adding device
I logged into a clean developer account and began following this guide. (Thank you much!)

When I was going through the steps to add a device, there were no locations or hubs listed.

I had to go to locations > and hubs to get this information to populate.

Asking the user to verify that their hub is listed will ensure it is available during the Add device steps.